refactor import to avoid circular import issue
