+++ toqito/channel_props/is_trace_preserving.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from toqito.channels import partial_trace
+from toqito.matrix_ops import partial_trace
 from toqito.matrix_props import is_identity
