+++ toqito/channels/realignment.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from toqito.channels import partial_transpose
+from toqito.matrix_ops import partial_transpose
 from toqito.perms import swap
