+++ toqito/matrix_ops/partial_trace.py
@@ -7,6 +7,8 @@ from cvxpy.expressions.variable import Variable
 from toqito.helper import expr_as_np_array, np_array_as_expr
 from toqito.perms import permute_systems
 
+# from toqito.channels import kraus_to_choi
+
 
 def partial_trace(
     input_mat: np.ndarray | Variable,
@@ -55,7 +57,7 @@ def partial_trace(
     .. jupyter-execute::
 
      import numpy as np
-     from toqito.channels import partial_trace
+     from toqito.matrix_ops import partial_trace
 
      test_input_mat = np.arange(1, 17).reshape(4, 4)
 
@@ -75,7 +77,7 @@ def partial_trace(
     .. jupyter-execute::
 
      import numpy as np
-     from toqito.channels import partial_trace
+     from toqito.matrix_ops import partial_trace
 
      test_input_mat = np.arange(1, 17).reshape(4, 4)
 
@@ -87,7 +89,7 @@ def partial_trace(
     .. jupyter-execute::
 
      import numpy as np
-     from toqito.channels import partial_trace
+     from toqito.matrix_ops import partial_trace
 
      test_input_mat = np.arange(1, 257).reshape(16, 16)
      test_input_mat
@@ -99,7 +101,7 @@ def partial_trace(
     .. jupyter-execute::
 
      import numpy as np
-     from toqito.channels import partial_trace
+     from toqito.matrix_ops import partial_trace
 
      partial_trace(test_input_mat, [0, 2], [2, 2, 2, 2])
