+++ toqito/state_opt/optimal_clone.py
@@ -3,8 +3,7 @@
 import cvxpy
 import numpy as np
 
-from toqito.channels import partial_trace
-from toqito.matrix_ops import tensor
+from toqito.matrix_ops import partial_trace, tensor
 from toqito.perms import permutation_operator
