+++ toqito/state_props/has_symmetric_extension.py
@@ -41,7 +41,7 @@ def has_symmetric_extension(
 
         import numpy as np
         from toqito.state_props import has_symmetric_extension
-        from toqito.channels import partial_trace
+        from toqito.matrix_ops import partial_trace
         rho = np.array([[1, 0, 0, -1], [0, 1, 1/2, 0], [0, 1/2, 1, 0], [-1, 0, 0, 1]])
         # Show the closed-form equation holds
         np.trace(np.linalg.matrix_power(partial_trace(rho, 1), 2)) >= np.trace(rho**2) - 4 * np.sqrt(np.linalg.det(rho))
