+++ toqito/state_props/is_ppt.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 
-from toqito.channels import partial_transpose
+from toqito.matrix_ops import partial_transpose
 from toqito.matrix_props import is_positive_semidefinite
